feat: Consolidate events page to root URL
- Move /events page content to root page (/)
- Move EventCard and SourceTabs components to app/components/
- Update constants.ts to point EVENTS_PAGE_URL to root
- Fix badge text from "毎週金曜日更新" to "毎日自動更新"
- Update footer text to "Made by MAKOTO IWABUCHI"
