+++ src/server/constants.ts
@@ -5,5 +5,5 @@
 // Web application URL
 export const APP_URL = "https://beer-line-watcher.vercel.app";
 
-// Events page URL for LINE messages
-export const EVENTS_PAGE_URL = `${APP_URL}/events`;
+// Events page URL for LINE messages (now at root)
+export const EVENTS_PAGE_URL = APP_URL;
